Fixed an issue with iOS9/10 (#43)
* Fixed an issue with iOS9/10

Authenticator is never added to the map, making auth fail on 9/10 for me at least. 

Adding this solved the bug on my end, could be I was just missing something, but if not, this should be a welcome one for people also having to work with legacy devices.

* Moved the fix to a different file to make it clean

* Added to the other file...

working with GitHub and multiple files isn't ideal, and I don't want to clone it all over again and install the right IDE for this, sorry!